Moved BoatRental to controller and renamed it to BoatController
